Fix tests and move api in its own serving package
