+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/project/ProjectService.java
@@ -47,7 +47,7 @@ import io.hops.hopsworks.api.jobs.KafkaService;
 import io.hops.hopsworks.api.jupyter.JupyterService;
 import io.hops.hopsworks.api.pythonDeps.PythonDepsService;
 import io.hops.hopsworks.api.tensorflow.TensorBoardService;
-import io.hops.hopsworks.api.tensorflow.TfServingService;
+import io.hops.hopsworks.api.serving.TfServingService;
 import io.hops.hopsworks.api.util.JsonResponse;
 import io.hops.hopsworks.api.util.LocalFsService;
 import io.hops.hopsworks.common.constants.message.ResponseMessages;
