+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/serving/ServingConfResource.java
@@ -14,7 +14,7 @@
  * If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.hops.hopsworks.api.tensorflow;
+package io.hops.hopsworks.api.serving;
 
 import io.hops.hopsworks.api.filter.NoCacheResponse;
 import io.hops.hopsworks.common.serving.tf.TfServingController;
